Cluster test utils - raname e move
