+++ tests/src/main/java/com/graphaware/test/integration/util/HAClusterIntegrationTestUtils.java
@@ -13,26 +13,28 @@
  * the GNU General Public License along with this program.  If not, see
  * <http://www.gnu.org/licenses/>.
  */
-package com.graphaware.runtime.config;
+package com.graphaware.test.integration.util;
 
 import java.io.File;
 import java.util.HashMap;
 import java.util.Map;
 
 import org.neo4j.causalclustering.core.CausalClusteringSettings;
-import org.neo4j.causalclustering.core.CoreGraphDatabase;
+import org.neo4j.cluster.ClusterSettings;
 import org.neo4j.graphdb.GraphDatabaseService;
+import org.neo4j.kernel.ha.HaSettings;
+import org.neo4j.kernel.ha.HighlyAvailableGraphDatabase;
 import org.neo4j.kernel.impl.factory.GraphDatabaseFacadeFactory.Dependencies;
 
 /**
- * Utility for causal cluster core instances 
+ * Utility for HA cluster
  */
-public class CoreCausalClusterIntegrationTestUtils extends AbstractClusterIntegrationTestUtils {
+public class HAClusterIntegrationTestUtils extends AbstractClusterIntegrationTestUtils {
 
 	@Override
 	protected GraphDatabaseService newDatabaseInstance(File storeDir, Map<String, String> params,
 			Dependencies dependencies) {
-		return new CoreGraphDatabase(storeDir, params, dependencies);
+		return new HighlyAvailableGraphDatabase(storeDir, params, dependencies);
 	}
 
 	@Override
@@ -41,13 +43,10 @@ public class CoreCausalClusterIntegrationTestUtils extends AbstractClusterIntegr
 		
 		params.put(CausalClusteringSettings.expected_core_cluster_size.name(), String.valueOf(clusterSize));
 		
-		params.put(CausalClusteringSettings.initial_discovery_members.name(), buildDiscoveryAddresses(5000,clusterSize));
-		params.put(CausalClusteringSettings.discovery_listen_address.name(), "localhost:500"+i);
-		params.put(CausalClusteringSettings.transaction_listen_address.name(), "localhost:600"+i);
-		params.put(CausalClusteringSettings.raft_listen_address.name(), "localhost:700"+i);
-		
-		params.put("dbms.connector.bolt.enabled", "true");
-		params.put("dbms.connector.bolt.listen_address", "localhost:"+ String.valueOf( 7687+i ));
+		params.put(ClusterSettings.server_id.name(), String.valueOf(i));
+		params.put(HaSettings.ha_server.name(), "localhost:600"+i);//"ha.host.data"
+		params.put(ClusterSettings.cluster_server.name(), "localhost:500"+i);//"ha.host.coordination"
+		params.put(ClusterSettings.initial_hosts.name(), buildDiscoveryAddresses(5000,clusterSize));//ha.initial_hosts
 		
 		return params;
 	}
